changes made to move sitemap and rssfeeds into one file
